I dont know if it is a corrent choice, but i dont like all things around me expect the one.
someone said that we should do the best thing so that we can get success, but it is too hard, dont it?

the more pressure, the less i could get.

It seems that i will get nothing except failure, but how to define the success or failure? i dont know. and i have done not know them alot of time. maybe its time to fly from the window right on me